isolated read and write handlers in their own modules and imported the handlers into playlist main.py
